requirements.txt
gcsa    #google calendar simple api - simplify managing google calendar API
python-dotenv    #use enviroment variables